requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==2.0.1
idna==3.4
Pillow==9.3.0
playwright==1.28.0
pyee==9.0.4
requests==2.28.1
typing_extensions==4.4.0
urllib3==1.26.13
